requirements.txt
python-telegram-bot==20.5
fastapi>=0.109.2
uvicorn>=0.23.2
yt-dlp>=2025.8.1
requests>=2.31.0
mutagen>=1.46.0